requirements.txt
firebase-admin
pandas
python-dotenv